- Adding print options class to control formatting - Changing StringBuilder to Appendable for streaming support
