+++ kotlin-xml-builder/src/main/kotlin/org/redundent/kotlin/xml/Element.kt
@@ -1,0 +1,12 @@
+package org.redundent.kotlin.xml
+
+/**
+ * Base interface for all elements. You shouldn't have to interact with this interface directly.
+ * @author Jason Blackwell
+ */
+interface Element {
+	/**
+	 * This method handles creating the xml. Used internally
+	 */
+	fun render(builder: Appendable, indent: String, printOptions: PrintOptions)
+}
