+++ kotlin-xml-builder/src/main/kotlin/org/redundent/kotlin/xml/PrintOptions.kt
@@ -1,0 +1,6 @@
+package org.redundent.kotlin.xml
+
+class PrintOptions(
+		val pretty: Boolean = true,
+		val singleLineTextElements: Boolean = false,
+		val useSelfClosingTags: Boolean = true)
