+++ kotlin-xml-builder/src/test/resources/test-results/XmlBuilderTest/noSelfClosingTag.xml
@@ -1,0 +1,3 @@
+<root>
+	<element></element>
+</root>
