+++ kotlin-xml-builder/src/test/resources/test-results/XmlBuilderTest/singleLineTextElement.xml
@@ -1,0 +1,4 @@
+<root>
+	<element>Hello</element>
+	<otherElement>Test</otherElement>
+</root>
